+++ workspace_new/template/.formatter.exs
@@ -1,6 +1,6 @@
 # Used by "mix format"
 [
-  inputs: ["mix.exs", "config/*.exs"],
+  inputs: ["{mix,.formatter,.workspace}.exs", "config/*.exs"],
   # by default packages are expected under a packages directory but you
   # can add any other subdirectory
   subdirectories: ["packages/*"]