+++ workspace_new/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+* Include `.workspace.exs` and `.formatter.exs` in generated `.formatter.exs`
+
 ## [v0.1.0](https://github.com/sportradar/elixir-workspace/tree/workspace_new/v0.1.0) (2024-05-13)
 
 Initial release.